Initial commit
Allows server owners to place vehicles into one of four predefined whitelists in the config, which will then require the respective password in-game before players can enter these vehicles. Passwords are set in the server file. Please note, that for players to enter a vehicle they have not got the password to, such as a suspect getting into a Police Car, they will need to be placed in the vehicle via an external script; this will likely be fixed in a future version.